refactor: LottoGenerator 인터페이스 사용을 위해 분리.
